Fjern utdaterte endepunkter
