+++ apps/api/src/main/kotlin/no/nav/helsearbeidsgiver/inntektsmelding/api/hentforespoersel/HentForespoerselRoute.kt
@@ -4,17 +4,14 @@ import io.ktor.http.HttpStatusCode
 import io.ktor.server.routing.Route
 import io.ktor.server.routing.RoutingRequest
 import io.ktor.server.routing.get
-import io.ktor.server.routing.post
 import kotlinx.serialization.builtins.serializer
 import kotlinx.serialization.json.JsonElement
 import no.nav.helsearbeidsgiver.felles.EventName
 import no.nav.helsearbeidsgiver.felles.Key
 import no.nav.helsearbeidsgiver.felles.Tekst
 import no.nav.helsearbeidsgiver.felles.domene.HentForespoerselResultat
-import no.nav.helsearbeidsgiver.felles.domene.ResultJson
 import no.nav.helsearbeidsgiver.felles.json.toJson
 import no.nav.helsearbeidsgiver.felles.kafka.Producer
-import no.nav.helsearbeidsgiver.felles.metrics.Metrics
 import no.nav.helsearbeidsgiver.felles.rapidsrivers.redis.RedisConnection
 import no.nav.helsearbeidsgiver.felles.rapidsrivers.redis.RedisPrefix
 import no.nav.helsearbeidsgiver.felles.rapidsrivers.redis.RedisStore
@@ -27,13 +24,8 @@ import no.nav.helsearbeidsgiver.inntektsmelding.api.auth.ManglerAltinnRettighete
 import no.nav.helsearbeidsgiver.inntektsmelding.api.auth.Tilgangskontroll
 import no.nav.helsearbeidsgiver.inntektsmelding.api.auth.lesFnrFraAuthToken
 import no.nav.helsearbeidsgiver.inntektsmelding.api.logger
-import no.nav.helsearbeidsgiver.inntektsmelding.api.response.RedisTimeoutResponse
 import no.nav.helsearbeidsgiver.inntektsmelding.api.sikkerLogger
-import no.nav.helsearbeidsgiver.inntektsmelding.api.utils.receive
 import no.nav.helsearbeidsgiver.inntektsmelding.api.utils.respond
-import no.nav.helsearbeidsgiver.inntektsmelding.api.utils.respondBadRequest
-import no.nav.helsearbeidsgiver.inntektsmelding.api.utils.respondForbidden
-import no.nav.helsearbeidsgiver.inntektsmelding.api.utils.respondInternalServerError
 import no.nav.helsearbeidsgiver.utils.json.fromJson
 import no.nav.helsearbeidsgiver.utils.json.toJson
 import no.nav.helsearbeidsgiver.utils.json.toPretty
@@ -47,61 +39,6 @@ fun Route.hentForespoersel(
     redisConnection: RedisConnection,
 ) {
     val redisPoller = RedisStore(redisConnection, RedisPrefix.HentForespoersel).let(::RedisPoller)
-
-    post(Routes.HENT_FORESPOERSEL_GAMMEL) {
-        val kontekstId = UUID.randomUUID()
-
-        Metrics.hentForespoerselEndpoint.recordTime(Route::hentForespoersel) {
-            runCatching {
-                receive(HentForespoerselRequest.serializer())
-            }.onSuccess { request ->
-                logger.info("Henter data for uuid: ${request.uuid}")
-                try {
-                    tilgangskontroll.validerTilgangTilForespoersel(call.request, request.uuid)
-
-                    val arbeidsgiverFnr = call.request.lesFnrFraAuthToken()
-
-                    producer.sendRequestEvent(kontekstId, request.uuid, arbeidsgiverFnr)
-
-                    val resultatJson = redisPoller.hent(kontekstId)
-
-                    sikkerLogger.info("Hentet forespørsel: $resultatJson")
-
-                    val resultat = resultatJson.success?.fromJson(HentForespoerselResultat.serializer())
-                    if (resultat != null) {
-                        respond(HttpStatusCode.Created, resultat.toResponse(), HentForespoerselResponse.serializer())
-                    } else {
-                        val feilmelding = resultatJson.failure?.fromJson(String.serializer()) ?: "Teknisk feil, prøv igjen senere."
-                        val response =
-                            ResultJson(
-                                failure = feilmelding.toJson(),
-                            )
-                        respond(HttpStatusCode.ServiceUnavailable, response, ResultJson.serializer())
-                    }
-                } catch (e: ManglerAltinnRettigheterException) {
-                    val response =
-                        ResultJson(
-                            failure = "Du har ikke rettigheter for organisasjon.".toJson(),
-                        )
-                    respondForbidden(response, ResultJson.serializer())
-                } catch (_: RedisPollerTimeoutException) {
-                    logger.info("Fikk timeout for ${request.uuid}")
-                    val response =
-                        ResultJson(
-                            failure = RedisTimeoutResponse(request.uuid).toJson(RedisTimeoutResponse.serializer()),
-                        )
-                    respondInternalServerError(response, ResultJson.serializer())
-                }
-            }.onFailure {
-                logger.error("Klarte ikke lese request.", it)
-                val response =
-                    ResultJson(
-                        failure = "Mangler forespørsel-ID for å hente forespørsel.".toJson(),
-                    )
-                respondBadRequest(response, ResultJson.serializer())
-            }
-        }
-    }
 
     get(Routes.HENT_FORESPOERSEL) {
         val kontekstId = UUID.randomUUID()
