--- apps/api/src/main/kotlin/no/nav/helsearbeidsgiver/inntektsmelding/api/utils/FjernLedendeSlash.kt
@@ -1,9 +1,0 @@
-package no.nav.helsearbeidsgiver.inntektsmelding.api.utils
-
-// Midlertidig workaround, frontend leverer ?uuid=/<uuid> foreløpig
-fun fjernLedendeSlash(str: String): String {
-    if (str.isNotEmpty() and str.startsWith("/")) {
-        return str.substring(1)
-    }
-    return str
-}
